Loop analyzer kind of works; broken handling set-cell and add-cell-0
The function `analyze-loop` needs to handle `set-cell` correctly; I'm
thinking I will need to not map to values but to structs like

    (mult 7) or (abs 6)

to ensure that the `mults` get multiplied by the current cell, while
the `abs` get set to a particular value. This will take a little bit
of work.

As it stands, the native compiler doesn't begin to handle the new
`mult-block-0` instruction; I'm just working on getting mandel.b to
run with both `opt/zero-out` and `opt/basic-loop` enabled.